Fix not supported Windows path delimiter in input paths to make imagemin cross platform
imagemin relies on globby package which (at least as of time of this commit)
expects paths to use forward slash (Unix) and not backward slash (Windows).

To fix this issue, I've added additional step, converting filePaths to unix format.

Unix format (forward slash) delimiter is supported in Windows for the most of cases.
@see https://superuser.com/questions/176388/why-does-windows-use-backslashes-for-paths-and-unix-forward-slashes/176395\#176395